moved print methods out to new class
